requirements.txt

pandas>=2.0
numpy>=1.25
plotly>=5.18
streamlit>=1.36
python-dotenv>=1.0
pytest>=8.0
loguru>=0.7
pyyaml>=6.0
